cmd/dist: add JSON output support for some tests
This change adds a "-json" flag to "dist test" that causes dist to pass
"-json" to the "go test" commands it runs, when possible. The "-json"
flag for "go test" produces machine-readable JSON output, which can be
useful for CI.

Dist runs some tests without using "go test", or runs them in a way that
aren't compatible with "-json". The output of these tests is unchanged.

Updates #37486