Add backend code gen
Adds code generation that works with the final list of crypto backends
to add extra features and tie up loose ends in an easily maintained way.

Backend conflict error generator: creates files that generate
informative build errors when the backends aren't configured correctly.

"nobackend" build constraint generator: gathers the build constraints
for all the backends to create the "nobackend" constraint.

The generators also run during standard test runs in validation mode to
ensure the generated content is not out of date when backends are
added/removed or when their build constraints change.

To simplify patch maintenance, the generators are only run once, in this
patch. An alternative would be adding the generators in an earlier patch
and re-running the generators immediately for each backend patch. This
would make the direct connection between changes clearer, but make it
harder to update the generators and to deal with conflicts.

Use "go/bin/go generate crypto/internal/backend" after recently building
the repository to run the generators.